Changed from OutputWindow to Logger for messages
On Windows, using the itkOutputWindow class for messages results in
a pop up window that can quickly disappear when the process exits.
Or somethings that window just hangs.  I think it has something to
do with not having a Windows event loop.

Now it uses the itkLogger class for message output, and the messages
get sent to the command line window (if the process is launched from
  the command line).