Fix Python index test with empty comparison
This addresses a test failure with Python 3.7 due to a change in the
StopIteration exception handling from PEP 479.

This test case has a slice index which results in an empty image. The
iteration produces a StopIteration exception which with Python 3.7 is
converted into a Run-time exception:

Traceback (most recent call last):
  File
  "/home/circleci/SimpleITK/Testing/Unit/Python/ImageIndexingTest.py",
  line 87, in test_2d
    self.assertImageNDArrayEquals(img[-4:-1:-2,-2:-1:1],nda[-2:-1:1,-4:-1:-2])
  File
  "/home/circleci/SimpleITK/Testing/Unit/Python/ImageIndexingTest.py",
  line 35, in assertImageNDArrayEquals
    for p,n in zip(img, nda.flat):
RuntimeError: generator raised StopIteration