Improve test coverage for ImageViewer class.
The ImageViewer test now invokes the class's Execute method.
The greatly improves the coverage for the class's code.

Prior to calling Execute the display command is set to use
'cmake -E md5sum' which computes the md5sum on input image.
This avoids trying to pop up an ImageJ or Fiji viewer.