Add isysroot as require flag in non-cmake superbuild projects
With XCode 10, the system headers are no longer installed in the
normal /usr/incude paths, even when the commandline tools are
installed. The passage of the SDK root via isysroot is nearly
required.

This issue manifests itself during the PCRE configuration step where
the it is reported: "error: cannot run C compiled programs."