Changed the github links from directly pointing to the latest release.
The previous link led to a page which only contained information with
respect to the latest version/release. It is better to point to the
page where all releases appear and the user can search that
page. Otherwise they won't know where to go if by latest release, they
actually wanted the release and not latest development version.