Update copyright assignment to NumFOCUS
SimpleITK is a proud part of the Insight Software Consortium (ISC),
and maintains a mirror under the official Github ISC organization.

As part of the process of the Insight Software Consortium joining the
NumFOCUS non-profit corporation as a Fiscally Sponsored Project, all
assets, including intellectual property assets, are being transferred
to enable dissolution of the Insight Software Consortium non-profit
corporate legal entity.