Update JSON doxygen strings from ITK
This is the first update with ITKv5 Doxygen, and a new version of
Doxygen used to generate the XML documentation. Mostly the class level
strings have been updated. Many of the method level doxygen strings
appears not to be currently correct in ITK, so most of them are not
updated in the patch.