Add filter execute methods mapping rvalue inputs to inplace
Adds to image filter member functions with rvalue as the first
input. These methods set the ITK filter  "in place" member to true,
which enables the filter to take the image buffer of the input and
re-use it for the output of the filter. This can improve performance
and reduces memory usage.

Updates the standard ImageFilterTemplate and expand template
components to add add the new Execute member function.