Check for aliasing of image's pixel container
To add additional safety with image buffer aliasing, the pixel
container of the itk::Image is now checked to verify that is only has
one reference. In filter's execute method the filter needs to be
deleted before the check to ensure the filter has no internal
references to the buffer.