Remove specific FastMarchingImageFilter template
The general image filter template work, with code to convert the
SimpleITK trial seed points to the itk node representation. This was
also done in the FastMarchingImageFilter, now the
FastMarchingBaseImageFilter uses the same approach. This reduces the
maintenance burden on the template code.

This change also removed the AddTrialPoints and ClearTrialPoints methods.