Explained the CMakeLists.txt hierarchy
There are two CMakeLists.txt files, one in the root CppCMake
dir and another in the CppCMake/Source subdir.  The second one
is the plain one that illustrates how to build a SimpleITK C++
project.  The first, root dir, one is to build sitk_example.cxx
in the test process.

Also added copywrite notice to sitk_example.cxx