Use NPasteImageFilter to replace PasteImageFilter
The NPasteImageFilter adds support to paste an M dimensional image
into an N-D image where N>=M, and support pasting a constant into a
region as well. This filter comes from the ITK remote module
SimpleFilters module, and is internally replacing the
itk::PasteImageFilter.

The DestinationSkipAxes parameter is added to the sitk filter, so
support the pasting with different dimensions.