Add to Java interface Image::GetBufferAsByteBuffer
The methods is renamed from GetBufferAsVoid to return a Java
byte-buffer object to Java-level code. The ByteBuffer can be converted
to other buffer type with methods such as `asFloatBuffer` and
`asIntBuffer`. The buffer object does not maintain a reference to the
Image, so the sitk::Image object must persists. Other documented
caveat of reading an writing to the raw image buffer persist.