Only require ITKCommon when finding ITK package
Reduce the dependency requirements when finding ITK. All the available
ITK modules are still listed, but not loaded, so the introspection of
the ITK configuration can still be done.

The addressed a build issue in 3D Slicer when ITKVTK module is
configured, but not of VTK dependencies can be found automatically.