WIP: Return an sitk::Transform with ITK specific pimple object
The returned object from ImageRegistrationMethod::Execute when the
InitialTransform had inPlace=False, was an sitk::Transform holding a
itk::CompositeTransform pimple object. The composite transform then
held an optimized transform of the initial type.

While the return type for the method remains a sitk::Transform, the
object held in the pimple transform is now the optimized transform of
the initial type.