DOC: Adding example of using Python logging with SimpleITK.
ITK issues warnings for several situations which are relevant for
debugging. These warnings are written to stderr or stdout. This is
primarily relevant for registration where the warning that not enough
points are available for estimating the similarity metric is
issued. This example illustrates how to re-route the file descriptors
so that the warnings are written to a log file.