BUG: Test for reading a DICOM series with variable rescale slope.
The output pixel type of the image is determined by the rescale slope
type. If only looking at a single image from a series, the first
image,  the type may be too restrictive (unsigned int) when other
images have a float. The image data is then truncated (e.g. 4.5 is
mapped to 4).