Exclude some characters for temp file name
The ImageViewer class generated a temporary file name from the user
supplied window title.  Some characters need to be removed because
it can cause problems when writing the file.

Previously it only checked for spaces.  Now it also checks for '/',
'\', ':', single quote and double quote.

Also a test has been augmented to exercise this code.