DOC: Updating the documentation for a corner case.
When reading a series of images, if the caller does not specify
the pixel type for the returned image, it is deduced from the
first image in the series. If this results in a narrowing
conversion, the returned image is not valid. The solution is
to have the user explicitly specify the wider pixel type. We do
not modify the approach because it would significantly increase the
computational cost while the use case is very rare.

closes #1636