Adding constant assignement support
Marked the AssignImage as optional, when not provided and the ivar is
a nullptr, then the constant will be used.

The DualImageFilter template did not support the combination of
inplace and optional inputs. As the Masked pixel type list was just a
single type, uint8, the template was changed to the regular
ImageFilter without change the support image arguments.