ENH: Dependency between GetGDCMSeriesIDs and GetGDCMSeriesFileNames.
The GetGDCMSeriesFileNames supports a useSeriesDetails flag which
appends tag based information to the seriesUID. This "decorated"
seriesUID is then used to identify files as belonging to the
series. The GetGDCMSeriesIDs did not support the useSeriesDetails
flag, so always returned the non-decorated seriesUID which in
combination with the decorated seriesUID used by
GetGDCMSeriesFileNames returns no matching files. This commit adds
the flag to the GetGDCMSeriesIDs so that returns "decorated"
seriesUIDs when it is set to true and will work as expected with
GetGDCMSeriesFileNames when that is called with the flag set to true too.