BUG: Fixed bug in example, working with DICOM tags.
The Study Instance UID tag was given as 0020|000D which is fine for
setting the value in the metadata dictionary but not when comparing to
the contents of the dictionary after reading from file. ITK/GDCM
converts the hexadecimal values to a lower case string.

The Series Description tag, 0008|103e, is optional and may
not exist.

Changes address both issues.