Restore functions for static members like ImageSeriesReader_GetGDCMSeriesIDs
SWIG 4.1.0 change behavior to removing flattened static methods for
objects. The compatible behavior can be restored by adding
`-flatstaticmethod` argument to SWIG. The
SimpleITK_PYTHON_FLATSTATICMETHOD CMake variable has been added to
control the usage of this flag. It is currently enabled by default and
is planned to default to OFF in future releases.