Restore and depricate MaskImageFilter support for mask input types
Instead of throwing an exception, implicitly convert the input mask to
a image of uint8. Also print a warning message that this behavior is
deprecated.

Add TypeListHasPixelIDValue and fix types in MaskImageFilter

Added to method, and used it the only cast integer pixel types for the
mask input image. This matches the previous behavior of the filter.