Remove signed label image support from label intensity filter
Only support unsigned label input images for the
LabelIntensityStatisticsImageFilter. This filter is a dual dispatch
filter with intensity image input. The generated output object file
for this filter was by far the largest. On a MAC ARM system this
reduced the size of the object file from 27MB to 14MB.