Update docker based package builds to use external source
Use the SimpleITK source code cloned by the action, and not internal
to the docker image.

Also remove the Python_EXECUTABLE variable being explicitly set during
transition of Python virtual environment.

Set HOME to the /tmp directory to avoid writing in the root of the
file system for user configuration.