COMP: address issue with recursive_directory_iterator on OSX CI.
The 'recursive_directory_iterator' used in the example is unavailable
on the OSX CI (error message says this iterator was introduced in
macOS 10.15). As this is not a key feature of the example, it was
simplified and no longer uses the iterator. Now, a given DICOM file is
converted to the specified output format.