Use a singleton for function factory in coded filter
Replace the member varialbe for the member function factory with a
function factory method which returns a singleton of for the function
factory. The singleton is now shared between instance and in
constant. This reduces the contruction time for each execution of a
filter.

Additional, the initialization of some member objects has been moved
to inline initialization.