Replace external project GTest with FetchContent Google Test
Removed the Superbuild GTest external project and replaced with
FetchContent in the SimpleITK core build. The superbuild only
downloaded the source and the GTest source was added as a
subdirectory. The FetchContent directly does this.

Note: To use an installed version of GTest/Google Test GOOGLETEST_ROOT
CMake variable should be set, previously GTEST_ROOT was expected to
point to googletest source code.